Implement numpy manager
